Split out deprecated stuff
